Handle certain NULL query results
Previously, if the Keen API responded with a result having a null entry,
parsing the response would result in a javascript error when this
project checks to see if the result needs to be cast to a Number.

This checks a few possible places were the API might return null, and does
not attempt to cast in these cases. This means null gets passed to the
chart, but in practice this hasn't been an issue. If it is, it could be
handled at the server level or the charting level, but at least this
project won't crash on it.